Revert "[SBG-583] Changed RetailStoreMenuItemOptionValue 'default' to bool from int. All tests pass."
This reverts commit 2b11ef8ab6c2b7d9e86db512fa830f1160eb9e1b.

+++ SnappyV2/Models/RetailStoreMenu/RetailStoreMenu+CoreData.swift
@@ -357,7 +357,7 @@ extension RetailStoreMenuItemOptionValue {
             id: Int(managedObject.id),
             name: managedObject.name ?? "",
             extraCost: managedObject.extraCost,
-            default: managedObject.defaultSelection,
+            default: Int(managedObject.defaultSelection),
             sizeExtraCost: sizeCosts
         )
         
@@ -372,7 +372,7 @@ extension RetailStoreMenuItemOptionValue {
         value.id = Int64(id)
         value.name = name
         value.extraCost = extraCost
-        value.defaultSelection = `default`
+        value.defaultSelection = Int16(`default`)
         
         if let sizeExtraCosts = sizeExtraCost {
             value.sizeCosts = NSOrderedSet(array: sizeExtraCosts.compactMap({ sizeExtraCost -> RetailStoreMenuItemOptionValueSizeCostMO? in
